@@ -21,8 +21,9 @@
     "@angular/platform-browser-dynamic": "^4.0.0",
     "@angular/router": "^4.0.0",
     "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^2.0.3",
-    "@ngrx/store": "^2.2.2",
+    "@ngrx/effects": "^4.0.3",
+    "@ngrx/store": "^4.0.0",
+    "@ngrx/store-devtools": "^4.0.0",
     "@types/lodash": "^4.14.67",
     "core-js": "^2.4.1",
     "lodash": "^4.17.4",
@@ -32,8 +33,8 @@
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.2",
-    "@angular/compiler-cli": "^4.0.0",
+    "@angular/cli": "1.3.2",
+    "@angular/compiler-cli": "^4.3.0",
     "@types/jasmine": "2.5.38",
     "@types/node": "~6.0.60",
     "codelyzer": "~2.0.0",
@@ -49,6 +50,6 @@
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "typescript": "~2.5.2"
   }
 }
