@@ -1,13 +1,13 @@
 import {ChangeDetectionStrategy, ChangeDetectorRef, Component, OnDestroy, OnInit} from '@angular/core';
 import {Store} from '@ngrx/store';
-import * as fromRoot from '../../reducers';
-import {Feature, FeatureId} from '../../../types/feature';
-import {EditFeature, StartEditFeature, UpdateEditedFeature, UpdateFeature} from '../../../actions/feature';
+import * as fromFeature from '../../reducers';
+import {Feature, FeatureId} from '../../types/feature';
+import {StartEditFeature, UpdateEditedFeature, UpdateFeature} from '../../actions/feature';
 import {ActivatedRoute} from '@angular/router';
 import {Subscription} from 'rxjs/Subscription';
 import 'rxjs/Rx';
 import {Observable} from 'rxjs/Observable';
-import {DetailsFormMode} from '../../../types/common';
+import {DetailsFormMode} from '../../../../types/common';
 
 @Component({
   selector: 'app-edit-feature-container',
@@ -23,12 +23,12 @@ export class EditFeatureContainerComponent implements OnInit, OnDestroy {
   editedDirty$: Observable<boolean>;
 
 
-  constructor(private _store: Store<fromRoot.State>, route: ActivatedRoute, cd: ChangeDetectorRef) {
+  constructor(private _store: Store<fromFeature.State>, route: ActivatedRoute, cd: ChangeDetectorRef) {
 
-    this.editedDirty$ = this._store.select(fromRoot.isEditedFeatureDirty);
-    this.formMode$ = this._store.select(fromRoot.getFeatureDetailsMode);
+    this.editedDirty$ = this._store.select(fromFeature.isEditedDirty);
+    this.formMode$ = this._store.select(fromFeature.getDetailsMode);
 
-    this._store.select(fromRoot.getEditedFeature)
+    this._store.select(fromFeature.getEdited)
       .subscribe((editedFeature) => {
         this.editedFeature = editedFeature;
         console.log('Edited feature: ', editedFeature);
@@ -54,7 +54,7 @@ export class EditFeatureContainerComponent implements OnInit, OnDestroy {
   }
 
   onStartEdit(featureId: FeatureId) {
-    this._store.dispatch(new EditFeature(featureId));
+    this._store.dispatch(new StartEditFeature(featureId));
   }
 
   ngOnDestroy() {
