@@ -1,36 +1,47 @@
-import * as fromFeature from './feature';
-import * as fromUser from './user';
+import * as fromFeature from '../feature/reducers/index';
+import * as fromUser from '../user/reducers/user';
 import {environment} from '../../environments/environment';
-import {ActionReducer, combineReducers} from '@ngrx/store';
-import {compose} from '@ngrx/core';
-import {storeFreeze} from 'ngrx-store-freeze';
-import {createSelector} from 'reselect';
+import {
+  ActionReducer, ActionReducerMap, createFeatureSelector, createSelector,
+  MetaReducer
+} from '@ngrx/store';
+// import {compose} from '@ngrx/core';
+// import {storeFreeze} from 'ngrx-store-freeze';
 
 export interface State {
-  feature: fromFeature.State;
-  user: fromUser.State;
+  // feature: fromFeature.State;
+  // user: fromUser.State;
+  x: string;
 }
 
-export const initialState: State = {
-  feature: fromFeature.initialState,
-  user: fromUser.initialState
-};
+export const getFeatureState = createFeatureSelector('feature');
+// export const getFeatureList = createSelector(getFeatureState, fromFeature.getFeatureList);
+// export const getEditedFeature = createSelector(getFeatureState, fromFeature.getEdited);
+// export const getSelectedFeatureId = createSelector(getFeatureState, fromFeature.getSelectedFeatureId);
+// export const getFeatureDetailsMode = createSelector(getFeatureState, fromFeature.getDetailsMode);
+// export const isEditedFeatureDirty = createSelector(getFeatureState, fromFeature.isEditedDirty);
 
-export const getFeatureState = (state: State) => state.feature;
-export const getFeatureList = createSelector(getFeatureState, fromFeature.getFeatureList);
-export const getEditedFeature = createSelector(getFeatureState, fromFeature.getEdited);
-export const getSelectedFeatureId = createSelector(getFeatureState, fromFeature.getSelectedFeatureId);
-export const getFeatureDetailsMode = createSelector(getFeatureState, fromFeature.getDetailsMode);
-export const isEditedFeatureDirty = createSelector(getFeatureState, fromFeature.isEditedDirty);
+export const reducers: ActionReducerMap<State> = {
+  // feature: fromFeature.reducer
+  //  , user: fromUser.reducer
+} as ActionReducerMap<State>;
 
-export const reducers = { feature: fromFeature.reducer};
-export const developmentReducer: ActionReducer<State> = compose(storeFreeze, combineReducers)(reducers);
-export const productionReducer: ActionReducer<State> = combineReducers(reducers);
 
-export function reducer(state: any, action: any) {
-  if (environment.production) {
-    return productionReducer(state, action);
-  } else {
-    return developmentReducer(state, action);
-  }
+// console.log all actions
+export function logger(reducer: ActionReducer<State>): ActionReducer<State> {
+  return function(state: State, action: any): State {
+    console.log('state', state);
+    console.log('action', action);
+
+    return reducer(state, action);
+  };
 }
+
+/**
+ * By default, @ngrx/store uses combineReducers with the reducer map to compose
+ * the root meta-reducer. To add more meta-reducers, provide an array of meta-reducers
+ * that will be composed to form the root meta-reducer.
+ */
+export const metaReducers: MetaReducer<State>[] = !environment.production
+  ? [logger]
+  : [];
