@@ -1,5 +1,5 @@
 import {Action} from '@ngrx/store';
-import {User} from '../types/user';
+import {User} from '../../../types/user';
 export const ADD_USER = 'Add User';
 
 export class AddUser implements Action {
@@ -7,3 +7,5 @@ export class AddUser implements Action {
   constructor(public payload: User) {
   }
 }
+
+export type Any = AddUser;
